+++ src/main/java/com/ddukbbegi/common/jwt/JwtFilter.java
@@ -1,13 +1,11 @@
-package com.ddukbbegi.common.auth;
+package com.ddukbbegi.common.jwt;
 
 import com.ddukbbegi.common.component.ResultCode;
-import com.ddukbbegi.common.exception.BusinessException;
 import jakarta.servlet.*;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.http.HttpStatus;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
@@ -15,8 +13,6 @@ import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.regex.Pattern;
